requirement.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.3
argcomplete==1.10.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.13.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==6.1.0
blinker==1.4
blis==0.7.11
Brotli==1.1.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
channels==4.1.0
chardet==3.0.4
charset-normalizer==3.3.1
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.3
comm==0.1.4
compressed-rtf==1.0.6
confection==0.1.3
configobj==5.0.8
configparser==7.0.0
constantly==23.10.4
cryptography==41.0.5
cymem==2.0.8
daphne==4.1.2
datefinder==0.7.3
Django==4.2.13
docx==0.2.4
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.15
etelemetry==0.3.1
exceptiongroup==1.1.3
extract-msg==0.28.7
filelock==3.15.1
fitz==0.0.1.dev2
frontend==0.0.3
guess-indian-gender==1.0.1
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==2.8
IMAPClient==2.1.0
importlib-resources==6.4.0
incremental==22.10.0
isodate==0.6.1
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language-data==1.2.0
looseversion==1.3.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.24.4
olefile==0.47
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20200517
pdfplumber==0.5.28
pgeocode==0.4.1
pillow==10.3.0
pocketsphinx==5.0.3
preshed==3.0.9
prov==2.0.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.4.2
pydantic-core==2.10.1
pydot==2.0.0
pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
PyMySQL==1.1.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf2==3.0.1
pypdfium2==4.30.0
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.2
python-pptx==0.6.23
pytils==0.4.1
pytz==2023.3.post1
pyxnat==1.6.2
rdflib
regex==2023.10.3
requests==2.32.3
rich==13.7.1
scipy==1.10.1
selenium==4.21.0
service-identity==24.1.0
shellingham==1.5.4
simplejson==3.19.2
six==1.12.0
smart-open==7.0.4
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
sqlparse==0.5.0
srsly==2.4.8
starlette==0.37.2
stop-words==2018.7.23
textblob==0.18.0.post0
thinc==8.2.4
tools==0.1.9
tqdm==4.66.4
traitlets==5.13.0
traits==6.3.2
trio==0.25.1
trio-websocket==0.11.1
txaio==23.1.1
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3
uvicorn==0.30.1
Wand==0.6.13
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
wrapt==1.16.0
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.0
zipp==3.19.2
zope.interface
